\name{DeconWK-package}
\alias{DeconWK-package}
\alias{DeconWK}
\docType{package}
\title{
Deconvolution by Weighted Kernels
}
\description{
This package contains code for density deconvolution using weighted
kernel estimators.  Type \sQuote{citation(\dQuote{DeconWK})} for details
of the implemented methods. 
}
\author{
Authors: Martin L Hazelton and Berwin A Turlach

Maintainer: Berwin A Turlach \email{berwin@maths.uwa.edu.au}
}
\references{
Hazelton, M.L. and Turlach, B.A. (2009). Nonparametric density
deconvolution by weighted kernel estimators,  Statistics and Computing
19(3): 217--228. \url{http://dx.doi.org/10.1007/s11222-008-9086-7}.  
}
\keyword{ package }
